sensor module sucessfully integrated in
The user controls feel a bit clunky though since you need to turn off the sensor enable and press reset every time you want to change the tempo. Since the reset is still glitched for the music player, this means that you may need to reload the program on the board every time you do this. Also, still need to convert the 16 b registers used for bram address to 8 b to allow htem to combine with the epp.